requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
